push-to-gar-docker: Always use buildx
We use the `gha` cache by default, and caches require using a different
driver. We currently only enable `buildx` when `platforms` is given, as
it is required for that type of build too. What we should do is always
enable it.

Do that, and add an input to allow the driver to be specified. That is
what `push-to-gar-docker` (the Artifact Registry companion of this
Docker Hub action) does, and the more similar we can have those be the
better.